requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
absl-py==1.0.0
    # via
    #   tensorboard
    #   tensorflow
astunparse==1.6.3
    # via tensorflow
cachetools==4.2.4
    # via google-auth
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.8
    # via requests
cycler==0.11.0
    # via matplotlib
flatbuffers==2.0
    # via tensorflow
fonttools==4.28.2
    # via matplotlib
gast==0.4.0
    # via tensorflow
google-auth==2.3.3
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-auth-oauthlib==0.4.6
    # via tensorboard
google-pasta==0.2.0
    # via tensorflow
graphviz==0.19
    # via -r requirements.in
grpcio==1.42.0
    # via
    #   tensorboard
    #   tensorflow
h5py==3.6.0
    # via tensorflow
idna==3.3
    # via requests
importlib-metadata==4.8.2
    # via markdown
keras==2.7.0
    # via tensorflow
keras-preprocessing==1.1.2
    # via tensorflow
kiwisolver==1.3.2
    # via matplotlib
libclang==12.0.0
    # via tensorflow
markdown==3.3.6
    # via tensorboard
matplotlib==3.5.0
    # via -r requirements.in
numpy==1.21.4
    # via
    #   -r requirements.in
    #   h5py
    #   keras-preprocessing
    #   matplotlib
    #   opt-einsum
    #   tensorboard
    #   tensorflow
oauthlib==3.1.1
    # via requests-oauthlib
opt-einsum==3.3.0
    # via tensorflow
packaging==21.3
    # via
    #   matplotlib
    #   setuptools-scm
pillow==8.4.0
    # via matplotlib
protobuf==3.19.1
    # via
    #   tensorboard
    #   tensorflow
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pydot==1.4.2
    # via -r requirements.in
pyparsing==3.0.6
    # via
    #   matplotlib
    #   packaging
    #   pydot
python-dateutil==2.8.2
    # via matplotlib
requests==2.26.0
    # via
    #   requests-oauthlib
    #   tensorboard
requests-oauthlib==1.3.0
    # via google-auth-oauthlib
rsa==4.8
    # via google-auth
setuptools-scm==6.3.2
    # via matplotlib
six==1.16.0
    # via
    #   absl-py
    #   astunparse
    #   google-auth
    #   google-pasta
    #   grpcio
    #   keras-preprocessing
    #   python-dateutil
    #   tensorflow
tensorboard==2.7.0
    # via tensorflow
tensorboard-data-server==0.6.1
    # via tensorboard
tensorboard-plugin-wit==1.8.0
    # via tensorboard
tensorflow==2.7.0
    # via -r requirements.in
tensorflow-estimator==2.7.0
    # via tensorflow
tensorflow-io-gcs-filesystem==0.22.0
    # via tensorflow
termcolor==1.1.0
    # via tensorflow
tomli==1.2.2
    # via setuptools-scm
typing-extensions==4.0.1
    # via tensorflow
urllib3==1.26.7
    # via requests
werkzeug==2.0.2
    # via tensorboard
wheel==0.37.0
    # via
    #   astunparse
    #   tensorboard
    #   tensorflow
wrapt==1.13.3
    # via tensorflow
zipp==3.6.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
